editremotes: select newly added remote
Previously after added a new remote, we would either reselect the
previously selected remote, or select the first remote in the list.
Since newly added remotes get added to the end of the list, the new
remote was therefore never selected.

This change makes the newly added remote selected in the list.

Closes #875
Signed-off-by: Tim Brown stimut@gmail.com