settings: fix issue with Recent repository count
The Recent repository count setting did not work as expected due to an off-by-one error.

If the setting was set to 4 for instance, then git-cola would keep track of the last 3 repositories (meaning the settings file would have the information on 3 repositories, one of which is the currently opened repository). This was because a nonstrict inequality was used instead of a strict inequality during a check.

With this fix, the settings file correctly now correctly keeps track of the last 4 repositories.

Signed-off-by: Victor Gambier victor.gambier@imt-atlantique.net