cmds: speed up showing binary untracked files
Apparently, Qt's text renderer is pretty slow when text includes
unprintable characters (code points below 32), or many different code
points, or if code points are not part of any fonts. This is possibly
due to font selection fallback handling or so.

Selecting an untracked binary file  can make the git-cola GUI
unresponsive for up to 1 or 2 seconds, even if the file size is
well below the default read limit of 2 KB.

Fix this by checking for errors during UTF-8 conversion. If the file
appears to be not UTF-8, treat the file as 7-bit ASCII. Code points
above 127 are replaced with the unicode replacement character (by
Python's decode()), and code points below 32 are manually replaced with
a '.'.

Although the encoding is normally UTF-8 (core.ENCODING), this also
respects different, user-configurable file encodings.

Signed-off-by: wm4 <wm4@nowhere>