git-cola-sequence-editor: move git-xbase to the bin/ directory
Migrate git-xbase out of the private share/git-cola/bin/ area and
install it into the general-purposes bin/ directory so that it
can be easily reused by users and external tools.

The distutils extras/ modules are no longer needed now that we
no longer rely on monkey-patching distutils to handle the
private bin/ directory.

Related-to: #181
Signed-off-by: David Aguilar <davvid@gmail.com>