setup.py: handle --use-env-python for Python3
Python3 requires that regex patterns that match bytes must be specified
using a bytes pattern.  The first_line_re is used to detect the Python
shebang lines, and Python3 reads files as bytes, so special-case the
pattern used on Python3.

Python2 accepts the unicode r"string literal" pattern when matching
against bytes.  Leave that code path as-is.

Closes #1102
Reported-by: David Roman <davidroman96@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>