main: register --help-commands on the main parser
Newer version of argparse on Pyhton 3.8 (and possibly earlier)
require us to register --help-commands on the main parser otherwise
an error message is reported when running "git cola --help-commands".

Add a new add_help_options() helper function and use it to register
the "--help-commands" flag on both the main parser and the cola
sub-parser.

Closes #1133
Reported-by: Fang Pengfei <pffang@vip.qq.com>
Signed-off-by: David Aguilar <davvid@gmail.com>