Diff: Allow partial stage/revert for untracked file
Currently an untracked file needs to be staged first
and then unwanted lines must be unstaged.
This change will allow "Stage selected line" and
"Revert selected lines..." to be used on untracked file.

The logic will be as followed:
When the user select "Stage selected line" on an untracked
file, these selected line will be added to staging, and
the rest of file will then be listed under modified.
When the user select "Revert selected lines.." on an
untracked file, those line will be deleted from the file

To generate the diff of the new file, dev null is used.
os.devnull is used instead of hard-coding '/dev/null/'.
This will ensure it will works in windows environment,
where 'nul' is needed instead of '/dev/null'

The downside of using this approach is the return code.
git diff will return `1` when used against new file or
file outside of index. Alternate check condition is used
by checking "new file mode" string from the output.
This should be enough to differentiate it from genuine
error such as filepath not found

Issue: #1084
Reported-By: mighty-d
Signed-off-by: mmargoliono