qtutils: guard against unhashable widget types
There was a short-lived bug in an earlier version of PyQt5 that
made widgets unhashable. Even though this bug was an oversight
and was quickly patched, it did temporarily break the ability to
use widgets as dict keys.

While this is not an issue in practice we can still make the
BlockSignals class robust to unhashable widgets.

Use id(widget) as the dict key rather using the widget as the
key directly.

Signed-off-by: David Aguilar <davvid@gmail.com>