qtutils: simplify the BlockSignals implementation
We don't need dicts to accomplish what BlockSignals needs to do.
We only need to iterate over the widgets and values in tandem.

We have a tuple of widgets. Build a list of values during
`__enter__`. `zip()` the `widgets` tuple and `values` list in
order to apply the values during `__exit__`.

Signed-off-by: David Aguilar <davvid@gmail.com>