models: emit a notification with model contents
The deferred nature of Qt notifications causes the about_to_update
signal to get handled *after* the model contents have been updated.

This makes it difficult to implement the selection restoration behavior
in the Status widget because we need access to the original state.

Emit a notification with the original state so that we can store and use
it in the Status widget.

Related-to: #1130 #1131
Signed-off-by: David Aguilar <davvid@gmail.com>