widgets: set a unique object name for all dock widgets
Qt requires that all dockwidgets must have unique object names in
order for serialization and restoration of Window settings to function.

We were incorrectly using a translated string as the object name,
which prevented the saved window settings from applying when the
language gets changed across cola invocations.

Fix this issue by passing a plain string to the create_dockwidget
creator function that will be used as the dockwidget object name.
Existing settings for English users will continue working as-is.
Settings for users in other languages may run into a one-time
bump when taking this update, but it should be an improvement
overall because the window settings are now properly saved
and restored across all sessions.

Closes #1071 #1161 #382
Helped-by: Bernd K <prof7bit@gmail.com>
Reported-by: Ｖ字龍(Vdragon) <Vdragon.Taiwan@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>