toolbar: Add option to set toolbar style
Qt natively supports different toolbar styles that users may wish to use
when customising the application.

By swapping "Show icon?" checkbox for a combo box, we can expose the
alternate styles. For backwards compatibility (and to prevent a crash),
the older "show_icons" configuration will continue to be saved.

The new "toolbar_style" key stores an integer, using variables to
reference the style in the code.

Signed-off-by: Luke Horwell <code@horwell.me>