dag: avoid duplicate notifications
Prior to this change, if a commit was selected in the DAG window by
clicking on a Commit item in the GraphView, two identical
diff.COMMITS_SELECTED notifications would be sent, once from
Commit.itemChange() for the previously selected Commit item, and again
from Commit.itemChange() for the newly selected Commit item.

Fix by no longer sending the notification from Commit.itemChange(), but
instead creating a new GraphView.selection_changed() method to send the
notification and connecting it to the scene.selectionChanged signal.

Signed-off-by: Daniel Harding <dharding@living180.net>