submodules: lazy load the submodules list
Do not run "git submoudle status" unless the submodules
widget is actually visible.

Lazy load the list of submodules. This list does not change
often. The Refresh button can be used to force the widget
to update if something changes on disk behind cola.

The expensive "git submodule status" call is now run once and
only once, and is not run on startup. The first call happens
when the submodules widget is first displayed.

Adding submodules interactively, switching repositories and
clicking the Refresh button will force the submodules list
to refresh.

Related-to: #1153
Signed-off-by: David Aguilar <davvid@gmail.com>