requirements: add PyQt5 and start reworking the README for Python3
Drop references to Python2 from the README. Start preparing for
the upcoming distutils removal by trimming bits we do not need
and removing references to older dependencies.

We will also be upgrading to a newer QtPy, which no longer
supports PyQt4, so we can stop mentioning it in our documentation.

Related-to: #1201
Signed-off-by: David Aguilar <davvid@gmail.com>