utils: fix insecure usage of NamedTemporaryFile
In the tmp_filename() function, a temporary file was being created by
tempfile.NamedTemporaryFile() and then immediately closed.  Because
no delete argument was passed to tempfile.NamedTemporaryFile(), this
meant the file was deleted as soon as it was closed.

Then when a caller of tmp_filename() opened a file using the file name
returned by tmp_filename(), it would actually be creating a new file
that did not have the same secure permissions as the file created by
tempfile.NamedTemporaryFile().

Fix this by passing delete=False to tempfile.NamedTemporaryFile().

Signed-off-by: Daniel Harding <dharding@living180.net>