cola: migrate the build system to setuptools
distutils is being removed in the near future. Prepare for this
eventuality by migrating to a pyproject.toml, setup.cfg and
setuptools-based build system.

Even though we advertise that the build system is Python3-only,
we still include compat try/except fallbacks that fallback
to distutils so that someone on Python2 could actually
install Git Cola if they needed to do so.

The bin/ wrapper scripts are still provided for convenience,
but they are much simpler now that they no longer need to
concern themselves with the actual details of setting
up the sys.path for installed cola distributions.

We are now using entry_points, so the bin/ wrappers are
effectively for developer convenience only. This allows
us to vastly simplify them while also advancing the
state of the build system by migraing to setup.cfg.

Closes #1122
Closes #1201
Signed-off-by: David Aguilar <davvid@gmail.com>