i18n: eliminate the need for .mo files
python setup.py build is deprecated, and we are relying on
it to generate the .mo files that we use for translations.

Eliminate the dependency on .mo files and related tooling by
by using polib to parse the .po files directly.

Move the po/*.po files to cola/i18n/*.po.
Implement the i18n module using polib.

Related-to: #1201