prefs: consult $VISUAL on edit when gui.editor is not set, too
Per our docs:
> The environment variable `$VISUAL` is consulted when no editor has
> been configured.

`$VISUAL` was only ever used as the default for the configuration
parameter in the preferences UI, i.e. it was not consulted when invoking
the editor (e.g. without ever going to the settings). In that sense, the
behavior arguably didn't match the doc. Change to consult it in both.