fix(prefs): do not write `cola.spellcheck` config on startup
When the `cola.spellcheck` config item has not been previously set,
simply starting up the app -- no need to actually do anything at all --
ends up writing its default value to the git config. This is mostly
harmless, but also unnecessary clutter of the git config with defaults.

Fix by taking the built in default into account when checking if it
should be written, not only the current configured value (which might be
unset).