cmds: allow the launcher to be a shell string
Right now the launcher can only be a single command name.
Allow the launcher to be specified as a shell string so that
we could possibly use other commands in the future that might
be invocations with more than one argument.

This would allow "gio open", but we're not going to change
our use of "xdg-open". "xdg-open" is the standard.

Related-to: #1245
Related-to: https://www.reddit.com/r/archlinux/comments/8mniks/comment/dzp1yd5
Signed-off-by: David Aguilar <davvid@gmail.com>