models: avoid resetting state in set_worktree()
During startup it is possible to have a Git object that already has its
worktree set to the correct value before set_worktree() is called.

This is because the worktree defaults to the current directory so we may
already have the correct values in our config reader.

Avoid resetting the config reader when the worktree is already
configured to the same directory.

This eliminates two additional calls to "git config" during startup.

Related-to: #499 #1153
Signed-off-by: David Aguilar <davvid@gmail.com>