status: ensure that untracked images can be staged
Enhance the tracking of modes to better differentiate between untracked
files that we can't diff (eg. binary files), untracked diffable files
and when we are displaying arbitrary text.

This allows us to better handle image files by being able to track when
files can be staged vs. files that can be partially staged.

Update the context menus for the status and diff widgets to be aware of
this distinction. This ensures that image files have a "Stage" entry in
their context menu.

Closes #1265
Reported-by: Gianni Lerro @glerroo via github.com
Signed-off-by: David Aguilar <davvid@gmail.com>