gitcfg: add support for even older git versions
Bring the minimum Git version back down again by adding a config reader
that works with even older versions than the one implemented for #1264.

This reader should be compatible with Git >= 1.7.10, which is
sufficiently old to avoid affecting any users.

While we do encourage users to upgrade to newer Git versions, we still
support older Git versions so that cola can be used in tandem with the
widest range of Git versions.

Related-to: #1264
Signed-off-by: David Aguilar <davvid@gmail.com>