branch: minimize refreshes and disable animations
The UI flickers due to animation being enabled on the tree widget.
Disable animation when rebuilding the tree to avoid flickers.

Rework the strategy used to defer updating of the branches widget so
that the refresh() routine is only called once during startup.

The initial call to refresh() will now do nothing because we now detect
that the model has never notified us of updates. This makes it so that
the showEvent() only calls refresh() when the widget is initially hidden
and then becomes shown.

Otherwise, the showEvent() that happens during initialization of the
widget is effectively a no-op. It is ignored because the model has not
yet sent any update() notifications.

The widget is then refreshed as soon as the model sends an update()
event.

Closes #1221
Signed-off-by: David Aguilar <davvid@gmail.com>