hotkeys: avoid hotkeys for "D" and "T"
Some configuration can cause the menu hotkeys to register in the context
of the commit message editor.

The hotkeys we put on the menu actions were only put there for aesthetic
purposes. Doing so make the hotkey text appear alongside the menu text
so that users can more easily discover the hotkeys.

Avoid registering these hotkeys to avoid having these hotkeys activated
when interacting with the text widgets. We'll lose the menu annotations,
but it should avoid issues observed on Debian/Linux with Qt 5.12.

Closes #1270
Reported-by: @lutra on github.com
Signed-off-by: David Aguilar <davvid@gmail.com>