main.py: Add repository search quick switcher
Add quick switcher for repository with Switcher class to main widget.
Quick switcher usually mapped to the shortcut ctrl/command+p, but it is
mapped for Pull action, so mapped this switcher to Alt+p.
By pressing the shortcut, the popup on top of the main widget will be showed.

Signed-off-by: nakanoi <nakanoinryota@gmail.com>