theme: apply the correct styling to the leading paths widget
Avoid issues with the leading paths widget going away by making it a
short-lived widget that is parented to the QMenu in which it operates.

Store the spinbox value in a persistent variable now that the widget is
transient. This is needed to retain the value, otherwise the value will
reset everytime the menu is recreated.

Signed-off-by: David Aguilar <davvid@gmail.com>