diffparse: refactor diff parsing
In the Patch.parse() classmethod, instead of initializing a _DiffHunk
with only the hunk header line and then appending the remaining lines
after it is initialized, use itertools.groupby() to capture all
the hunk's lines together so that they can be passed when initializing
the _DiffHunk.  While this takes slightly more code I feel it is cleaner
because now there are never _DiffHunk objects in an incomplete state.

Signed-off-by: Daniel Harding <dharding@living180.net>