diff: support editing patches before staging etc.
Add additional popup menu items such as "Edit Diff Hunk to Stage..." or
"Edit Selection to Stage..." (with similar items for Unstage and
Revert).  These items will save the appropriate patch to a temporary
file and launch the user's editor to allow the user to manually edit the
patch before applying.  Once the editor exits, the updated patch will be
loaded from the temporary file and applied as normal.

Closes: #794
Suggested-by: Kai Krakow <hurikhan77@gmail.com>
Suggested-by: Tyler Shellberg <tyler.shellberg@protonmail.com>
Signed-off-by: Daniel Harding <dharding@living180.net>