diff: use "remove" icon for Unstage menu item
Prior to this commit, the Unstage menu item in the DiffEditor context
menu for a staged file would have an "add" icon ("+") even though the
"Unstage Diff Hunk" menu item right below it has a "remove" icon ("-").
Update the code to dynamically switch the icon for the .stage_or_unstage
menu item between "add" or "remove" depending on whether the file is
modified or staged.

Signed-off-by: Daniel Harding <dharding@living180.net>