diff: fix the right-click context menus
We were using the customContextMenuRequested signal to setup the base
class context menu. This affected downstream classes that use
contextMenuEvent() in order to build their context menus.

Use contextMenuEvent() in the base class as well. Expose a new
add_links_to_menu() method that we can use from both the diff view and
the base class in order to decorate menus with "Open [link]" actions.

Closes #1287
Reported-by: Daniel Harding <dharding@living180.net>
Signed-off-by: David Aguilar <davvid@gmail.com>