gitcfg: improve the terminal support
Extend the logic that is used to find a suitable terminal emulator for
the "Launch Terminal" action. We now check for a wider assortment of
shells when `cola.terminal` is not configured.

Users can always configure this value if the priority list in this
modules is not to their liking.

Update how gnome-terminal is executed to avoid the
deprecated "-e" option. Use "--" for mate-terminal and gnome-terminal.

Related-to: #1223
Signed-off-by: David Aguilar <davvid@gmail.com>