diff: use a block cursor for better usability
The block cursor is a better cursor for the diff editor because it is
easier to see. #889 got rid of the block cursor, but that issue was
mostly focused on the Commit Message Editor, not the diff editor.

The diff editor is more cursor-centric, and its contents are
read-only, so the use of a more visible cursor is a better default.

Provide "git config --global cola.blockcursor false" and a
corresponding Preferences setting to disable the block cursor
in case there are users that prefer the original cursor style.

Related-to: #889
Signed-off-by: David Aguilar <davvid@gmail.com>