widgets: consolidate context menu handling for text widgets
Add a reusable create_context_menu() and context_menu_event()
implementation to the base extension classes so that we have a uniform
and consistent way to add actions to context menus.

A self.menu_actions field is now used for adding entries to the
beginning of the context menu. This eliminates the need to implement any
functions in the implementation classes for Grep and DiffTextEdit.

Signed-off-by: David Aguilar <davvid@gmail.com>