GUI: Stop making newly-created buttons immediately visible
On Windows, making a newly-created button visible causes
a new window to open containing only the new button.

Testing shows that the buttons will be visible
in whatever dock they are ultimately assigned to
when the main git-cola window opens
even if `.setVisible(True)` is not called first.

Therefore, the default `visible` value is changed to `None`,
and `.setVisible` will only be called if a bool value is specified.

Fixes #1329