commitmsg: perform commits asynchronously
Run "git commit" in the background so that we can provide feedback while
a commit is in progress.

Replace the commit button with a mini progress bar for the duration of
the "git commit" action.

Move the Interaction.command() calls to the editor from the
Commit command to prevent Qt widgets from being created from a
background thread (which crashes Qt).

Closes: #1320
Suggested-by: Peter Justin <peter@peterjustin.com>
Signed-off-by: David Aguilar <davvid@gmail.com>