i18n: restore most of the translations
We are going to stay on the pre-existing track of translating as much as
we can.

Only a few micro changes were done to the "stash" translations for
French after consulting with a native speaker.

The original issue reported in #1343 was centered around a specific
language, so a follow-up commit will adjust the Push and Pull buttons to
include a tooltip that clarifies their purposes.

Related-to: #1160 #1343 #1351
Signed-off-by: David Aguilar <davvid@gmail.com>