cmds: prevent StageUntracked from staging all files
When the list of files to stage is empty then the Stage command, which
is a base class of the StageUntracked command, will fallback to staging
all (modified) files.

This behavior is unwanted when dealing with untracked files. This is
doubly-so when the user has `gui.displayuntracked = false` configured
because the untracked file list will always be empty.

Closes: #1371
Reported-by: @pkzc via github.com
Signed-off-by: David Aguilar <davvid@gmail.com>