sequenceditor: ensure that tasks are stopped in all situations
Make sure that we stop running tasks when clicking the Cancel button,
when clicking the (X) window close button and when using the Ctrl-Q
quit hotkey.

We will only save the instruction sheet and return exit status zero
when the Rebase button is pressed. Exit status 1 is returned in all
other cases.

Signed-off-by: David Aguilar <davvid@gmail.com>