sequenceeditor: refine how multi-select actions are handled
We were previously toggling items individually. This is fine most of the
time but when using multi-select it is preferable to have the entire
selection resolve to a consistent value across all selected items.

Make it so that the "enabled" checkbox and the remarks calculate whether
its state should change based on the selection.

If more items are enabled then all items will become enabled. If more
items are disabled then all items will become disabled. Special-case
having all items enabled or disabled and toggle all items.

Related-to: #1375
Signed-off-by: David Aguilar <davvid@gmail.com>