polib: use a system-provided polib when available
Demote the cola.polib module to be a fallback module that is only
used when a system-provided polib is not available. We now attempt
to import the real polib instead of preferring our vendored module.
This sets the state for a the future where we might remove this
module completely.

Signed-off-by: David Aguilar <davvid@gmail.com>