push: set the remote branch field when selecting a local branch
Keep the behavior of the "Push" dialog closer to the original behavior
where selecting the local branch would implicitly push that remote
branch to the same-named remote branch.

We now clear selections from the remote branch list when a local branch
is clicked. But, we also pre-populate the "Remote Branch" field so that
it's clear that the selected branch will be pushed to the remote.

We can still push a disjoint set of branches by explicitly selecting
a remote branch, but that is more of an uncommon operation.

Closes: #729
Suggested-by: @pkzc on github.com
Signed-off-by: David Aguilar <davvid@gmail.com>