resources: search for commands sibling to the running script
When Cola is installed directly into C:\Python3.12 on Windows
then the correct path to git-cola-sequence-editor was not being used.

Look in the same director as the current sprint when attempting to
find git-cola-sequence-editor. This makes cola rely less on the
physical structure of the install tree by checking for locations
relative to itself.

Closes: #1385
Reported-by: Stanislaw Halik <sthalik@misaki.pl>
Signed-off-by: David Aguilar <davvid@gmail.com>