packaging: restore Python 3.7 support
version_file support is only available in setuptools_scm 8.0,
and older versions throw errors when pyproject.toml contains
configuration that the older versions do not understand.

This means that we cannot use version_file just yet if
we want to retain support for Python 3.7.

Restore the importlib.metadata version lookups alongside
the setuptools_scm version file support so that we can
swap over later once we are okay dropping support for 3.7.

Signed-off-by: David Aguilar <davvid@gmail.com>