models: fallback to GNOME for non-KDE environments
Ubuntu's XDG_CURRENT_DESKTOP value is set to "ubuntu:GNOME",
which causes the proxy autodetection to be a no-op on Ubuntu.

Debian/testing has XDG_CURRENT_DESKTOP set to "GNOME", so
we follow a similar pattern when checking for "KDE" ~
we will accept either "KDE" or a string that ends with ":KDE".

Rework the logic so that we fallback to using GNOME/gsettings
for non-KDE Linux desktop environments.

Closes: #1420
Helped-by: 林博仁(Buo-ren Lin) <buo.ren.lin@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>