version: update the PKG-INFO version when a Git Cola worktree is present
setuptools-scm records a version number when `pip install --editable .`
and other build commands are run (e.g. when generating wheels).

The version number is recorded in the `git_cola.egg-info/PKG-INFO` file,
and this file becomes out of date as soon as any commits are authored or
the worktree is updated to a newer commit.

Teach the developer wrapper scripts to update the `PKG-INFO` file when
its version number is out of sync with the source tree. When no
`PKG-INFO` file is present then we update the `cola.version.SCM_VERSION`
field with the updated version instead.

This ensures that users always see the most accurate version number
when running from a Git worktree.

Closes: #1425
Suggested-by: 林博仁(Buo-ren Lin) <buo.ren.lin@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>