dag: avoid "git log --reverse" to fix "git dag --follow"
When using "git log --follow" we end up with truncated output. This is
due to implementation details around how the --reverse option is
implemented in core git. It is arguably a bug.

We also pass `-#` to limit the output (e.g. "-2000"), and this further
hinders the ability to get the right results from git.

Change the commit reader to no longer rely on `--reverse`. We instead
reverse the commit lines ourselves after reading the output from git in
the default topo-order.

Closes: #1327
Reported-by: Michael Baumgartner @BaumiCoder on github
Signed-off-by: David Aguilar <davvid@gmail.com>