pyproject.toml: remove usage of setup.cfg
Stop relying on data_files to provide our hotkey html files. setuptools
does not support `data_files` via pyproject.toml so we have to use
package data instead.

For files that pip cannot handle, namely *.desktop and related Linux
desktop files, the garden.yaml and Makefile commands have been updated
to handle this responsibility.

Although redundant, we end up installing the hotkeys files in two
locations so that it's part of the documentation area and private to
cola's package data for use by the app itself. This effectively makes
the app independent of whether or not the docs have been installed.

Signed-off-by: David Aguilar <davvid@gmail.com>