commitmsg: improve the time slider behavior
Map the slider to time values between 00:00:00 and 23:59:59.

Do not update the date widget when the slider is changed. The time
slider will now control just the time fields.

This makes the behavior simpler and more predictable as the slider will
always represent the amount of time that has passed since midnight.

Add callbacks to keep the slider up to date in response to direct edits
to the time widget.

Switch to using a QTimeEdit instead of a QDateTimeEdit.

Related-to: #1429
Signed-off-by: David Aguilar <davvid@gmail.com>