commitmsg: streamline the interface
Move the summary and commit button up into the dockwidget title widget.
This consolidates space and further reduces visual clutter.

Rework how the cursor position label is displayed. Add rounded corners
and use stylesheets instead of the HTML snippets that were used
previously. This makes it easier to control the height of the label.

Signed-off-by: David Aguilar <davvid@gmail.com>