commitmsg: restore the original tab order
The tab order was removed in fffcc7e007103d1bdaec3416ba589a78c3bfaae1
when the summary and description fields were moved into different
widgets. Qt was emitting a warning that the call had no effect.

The correct way to fix this issue is to apply the tab order on the
widget's common parent, specifically the dock widget itself.

Change the scope of the setTabOrder call to the dockwidget, which is the
grandparent of both widgets, to restore the correct behavior when
tabbing from the summary field into the description field.

Closes: #1436
Reported-by: Stanislaw Halik <sthalik@misaki.pl>
Signed-off-by: David Aguilar <davvid@gmail.com>