dag: speedup and simplify how threads are managed
Remove the mutex that was used for pausing and resuming threads since we
no longer use these code paths. This avoids needing to take a look while
looping over each commit, which slightly things up.

Configure the the ReaderThread so that it can be terminated instead of
using a mutex and state variables to handle termination.

Signed-off-by: David Aguilar <davvid@gmail.com>