sequenceeditor: update the model during application startup
Recent changes to the RepoReader added logic to detect the "git init" state.
This had a side-effect of preventing the sequenceeditor from displaying
commits because its model is not updated for performance reasons.

Update the model during startup to allow the model to recognize that
we are not in the "git init" state. The next commit will add a special
mode that allows us to avoid having to update the model, which will
restore the original performance optimization.

Closes: #1442
Signed-off-by: David Aguilar <davvid@gmail.com>