sequenceeditor: restore the performance optimization during initialization
bb324cd1 (sequenceeditor: update the model during application startup, 2024-12-12)
fixed a regression by updating the model on startup, which is an
expensive operation that we can avoid.

Restore the ability to query for commits from the RepoReader without
needing to have an updated model. Set a state variable in the
sequenceeditor to opt-in to this behavior so that we can perform
the minimum number of Git operations during startup.

Related-to: #1442
Signed-off-by: David Aguilar <davvid@gmail.com>